+++ app/models/admin/research_article.rb
@@ -21,7 +21,7 @@ class Admin::ResearchArticle < ApplicationRecord
   # Concerns
   include Deletable
 
-  # Named scopes
+  # Scopes
   # Methods
   def to_param
     slug