+++ app/models/comment.rb
@@ -17,6 +17,6 @@ class Comment < ApplicationRecord
   # Concerns
   include Deletable
 
-  # Named scopes
+  # Scopes
   # Methods
 end