+++ app/models/engagement_response.rb
@@ -14,6 +14,6 @@ class EngagementResponse < ApplicationRecord
   # Concerns
   include Deletable
 
-  # Named scopes
+  # Scopes
   # Methods
 end