+++ app/models/reply.rb
@@ -12,7 +12,7 @@ class Reply < ApplicationRecord
   multisearchable against: [:description],
                   unless: :deleted_or_chapter_deleted?
 
-  # Named Scopes
+  # Scopes
   scope :points, -> { select('replies.*, COALESCE(SUM(reply_users.vote), 0)  points').joins('LEFT JOIN reply_users ON reply_users.reply_id = replies.id').group('replies.id') }
 
   # Model Validation